fix: landscape support
